requirements.txt
requests==2.31.0
sqlalchemy==2.0.23
pyyaml==6.0.1
python-dotenv==1.0.0
schedule==1.2.0
